refactor: deleted IssueModal to reword the UX. Visual fixes to navbar
