+++ package.json
@@ -161,7 +161,6 @@
     "overmind": "^28.0.1",
     "overmind-react": "^29.0.1",
     "prop-types": "^15.7.2",
-    "rc-slider": "^9.2.1",
     "react": "^17.0.2",
     "react-confirm-alert": "^2.5.0",
     "react-cool-inview": "^2.0.8",
