--- render/components/styles/rc-slider.css
@@ -1,5 +1,0 @@
-/* https://github.com/react-component/slider/issues/413 */
-/* bugfix: */
-.rc-slider-tooltip {
-  z-index: 99;
-}
