+++ render/views/__tests__/SummaryPage.spec.jsx
@@ -8,10 +8,10 @@ import configureStore from 'redux-mock-store';
 import MockAdapter from 'axios-mock-adapter';
 import { ThemeProvider } from 'styled-components';
 
-import { theme } from '../../../theme';
-import { SummaryPage } from '../../AppViewPages/SummaryPage';
-import issueActions from '../../../actions/issues.actions';
-import { initialize, getInstance, reset } from '../../../../common/request';
+import { theme } from '../../theme';
+import { SummaryPage } from '../SummaryPage';
+import issueActions from '../../actions/issues.actions';
+import { initialize, getInstance, reset } from '../../../common/request';
 
 const mockStore = configureStore([thunk]);
 const redmineEndpoint = 'redmine.test.test';
